rename .data-poster.use-leveldb to .use-db-storage, rename dataposter/leveldb package
