+++ arbnode/dataposter/dbstorage/storage.go
@@ -1,7 +1,7 @@
 // Copyright 2021-2023, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
-package leveldb
+package dbstorage
 
 import (
 	"bytes"
@@ -17,7 +17,7 @@ import (
 	"github.com/syndtr/goleveldb/leveldb"
 )
 
-// Storage implements leveldb based storage for batch poster.
+// Storage implements db based storage for batch poster.
 type Storage struct {
 	db     ethdb.Database
 	encDec storage.EncoderDecoderF
